requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
Pillow>=10.1.0,<10.2
drf-spectacular>=0.22.1,<0.23
